Add modular SQL input components with backward compatibility
